Split the eval function in Lean into multiple functions.
The idea of this was to prevent having to deal with optional arguments
in the Lean code, which means less boxing/unboxing.

The down side is we have more complex c++ logic to handle the
different cases. I don't want to shove that down the line to python,
otherwise the API will be more complex.

I'm not 100% sure on optimization, but I believe having more logic in
c++ is probably the best way to go for performance (however minimal it
ends up being).