Replace running computation in IO with FrontendM.
I'm modeling this work after Pantograph a bit. My original purpose was
to expose the cancelTk? object so I can ensure I cancel tasks. I
believe that token is passed down into CoreM, which is where the hard
work takes place.

I believe there's some more interesting stuff to be done here. Right
now we're essentially *getting around* the monads rather than actually
using them. This is because we want the "state" to be exported to
Python eventually. This makes sense for sitations where we want to do
some evaluation in lean, do something else in python, and then do some
more computation in Lean.

I am avoiding using a "server" model, where we have a lean process
which stays running, and then a python client which talks to it. I
wanted it to be a single function call to make it as braindead to use
as possible. However, there might be some benefits to running it as a
server, since we'd be able to use it over network. Maybe that will be
a secondary project which uses pyle.